fix: dynamo db
