+++ src/components/BackgroundCarousel.tsx
@@ -2,9 +2,7 @@ import { Box } from '@mui/material';
 import React from 'react';
 import { Carousel } from 'react-responsive-carousel';
 import 'react-responsive-carousel/lib/styles/carousel.min.css';
-
 import backgroud01 from '../assets/background.jpg'
-
 
 const BackgroundCarousel: React.FC = () => {
     const carouselPictures = [
